Xml comments for Policies
