+++ Source/Policies/Improvements/RecipeLocator.cs
@@ -8,16 +8,21 @@ using Dolittle.Types;
 
 namespace Policies.Improvements
 {
+    /// <inheritdoc />
     public class RecipeLocator : IRecipeLocator
     {
         readonly ITypeFinder _typeFinder;
-
+        /// <summary>
+        /// Instantiates an instance of <see cref="RecipeLocator" />
+        /// </summary>
+        /// <param name="typeFinder"></param>
         public RecipeLocator(ITypeFinder typeFinder)
         {
             _typeFinder = typeFinder;
         }
 
-        public IRecipe GetByName(RecipeType name)
+        /// <inheritdoc />
+        public IRecipe GetByType(RecipeType name)
         {
             var typeName = $"{typeof(RecipeLocator).Namespace}.Recipes.{name}";
             var type = _typeFinder.FindTypeByFullName(typeName);
