+++ Source/Policies/Improvements/StepContainerMustHaveCommand.cs
@@ -3,6 +3,7 @@
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 using System;
+using k8s.Models;
 
 namespace Policies.Improvements
 {
